+++ comptandye/src/test/java/fr/trendev/comptandye/security/controllers/jwt/JWTRevokedSetTest.java
@@ -7,7 +7,7 @@ package fr.trendev.comptandye.security.controllers.jwt;
 
 import static fr.trendev.comptandye.security.controllers.jwt.JWTManager.SHORT_VALID_PERIOD;
 import static fr.trendev.comptandye.security.controllers.jwt.JWTManager.SHORT_VALID_PERIOD_UNIT;
-import fr.trendev.comptandye.security.controllers.jwt.dto.firestore.MockJWTRevokedSetDTO;
+import fr.trendev.comptandye.security.controllers.jwt.dto.mock.MockJWTRevokedSetDTO;
 import fr.trendev.comptandye.security.entities.JWTRecord;
 import java.time.Instant;
 import static java.time.temporal.ChronoUnit.MINUTES;
