+++ comptandye/src/test/java/fr/trendev/comptandye/security/controllers/jwt/dto/mock/MockErrorsJWTWhiteMapDTO.java
@@ -5,7 +5,6 @@
  */
 package fr.trendev.comptandye.security.controllers.jwt.dto.mock;
 
-import fr.trendev.comptandye.security.controllers.jwt.dto.firestore.MockJWTWhiteMapDTO;
 import fr.trendev.comptandye.security.entities.JWTWhiteMapEntry;
 import java.util.Collections;
 import java.util.List;
