+++ comptandye/src/test/java/fr/trendev/comptandye/security/controllers/jwt/dto/mock/MockJWTRevokedSetDTO.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package fr.trendev.comptandye.security.controllers.jwt.dto.firestore;
+package fr.trendev.comptandye.security.controllers.jwt.dto.mock;
 
 import fr.trendev.comptandye.security.controllers.jwt.dto.JWTRevokedSetDTO;
 import fr.trendev.comptandye.security.entities.JWTRecord;
